refactor: reorganize query params module and improve value splitting logic
